feat: add feature upload image

+++ app/Controllers/PegawaiController.php
@@ -24,7 +24,8 @@ class PegawaiController extends BaseController
 
     public function show($id)
     {
-        // 
+        $data['pegawai'] = $this->modelPegawai->getPegawaiWithJabatanWhere($id);
+        return view('pegawai/show', $data);
     }
 
     public function create()
@@ -40,8 +41,14 @@ class PegawaiController extends BaseController
             'alamat' => $this->request->getPost('alamat'),
             'telepon' => $this->request->getPost('telepon'),
             'jabatan_id' => $this->request->getPost('jabatan_id'),
-
         ];
+        // menghandle foto
+        $fileFoto = $this->request->getFile('file_foto');
+        if ($fileFoto && $fileFoto->isValid() && !$fileFoto->hasMoved()) {
+            $namaFile = $fileFoto->getRandomName();
+            $fileFoto->move('uploads', $namaFile);
+            $data['foto_pegawai'] = $namaFile;
+        }
 
         $this->modelPegawai->save($data);
         return redirect()->to('pegawai');
@@ -63,13 +70,42 @@ class PegawaiController extends BaseController
             'jabatan_id' => $this->request->getPost('jabatan_id'),
 
         ];
+            // Ambil foto lama
+            $fotoLama = $this->request->getPost('foto_lama');
+
+        // menghandle foto
+        $fileFoto = $this->request->getFile('file_foto');
+        if ($fileFoto && $fileFoto->isValid() && !$fileFoto->hasMoved()) {
+            $namaFile = $fileFoto->getRandomName();
+            $fileFoto->move('uploads', $namaFile);
+            $data['foto_pegawai'] = $namaFile;
+
+            // hapus fotolama jika ada
+            $fotoLama = $this->request->getPost('foto_lama');
+            if (!empty($fotoLama)) {
+                $filePath = 'uploads/' . $fotoLama;
+                if (is_file($filePath)) {
+                    unlink($filePath);
+                }
+            }
+        }
 
         $this->modelPegawai->save($data);
         return redirect()->to('pegawai');
     }
     public function delete($id)
     {
-        $this->modelPegawai->delete($id);
+        // $this->modelPegawai->delete($id);
+        $pegawai = $this->modelPegawai->find($id);
+        if ($pegawai) {
+            if (!empty($pegawai->foto_pegawai)) {
+                $filePath = 'uploads/' . $pegawai->foto_pegawai;
+                if (is_file($filePath)) {
+                    unlink($filePath);
+                }
+            }
+            $this->modelPegawai->delete($id);
+        }
         return redirect()->to('pegawai');
     }
 }
